+++ MapDriveInator/history.md
@@ -1,5 +1,10 @@
 # History
 
+## v0.0.8
+
+* "normalization" command for absolute/relative links
+* now default "link type value" is taken form freeplane configuration properties
+
 ## v0.0.7
 
 * Test version allowing the use of RELATIVE and ABSOLUTE links